+++ Makefile
@@ -29,7 +29,7 @@ LIBFT			= $(LIBFT_DIR)$(LIBFT_NAME)
 
 NAME			= MiniRT
 CC				= cc
-FLAGS			= -Wall -Werror -Wextra -g3 -fstack-protector-all -fsanitize=undefined
+FLAGS			= -Wall -Werror -Wextra -g3 -fsanitize=undefined
 RM				= rm -f
 AR				= ar rcs
 NORM			= ./includes\
@@ -44,7 +44,7 @@ INCLUDE			= 	-I ./includes/\
 OBJ_DIR_F 		= obj/
 SRC_DIR_F		= src/
 SRC_FUNCTIONS_F	= main atof patterns textures textures2 colors2 color_calculation color_calculation_2 tuple_calculations tuple_calculations2 tuple_calculations_3 rays rays_2 rays_3 \
-					matrix_calcul matrix_deter_cofac hooks matrix_utils matrix_utils2 utils utils_2 init \
+					matrix_calcul threads matrix_deter_cofac hooks matrix_utils matrix_utils2 utils utils_2 init init2 init3 init4 init5 init6 init7 init8 \
 					parsing parsing_2 parsing_3 parsing_4 parsing_5 win get_next_line get_next_line_utils light_lst cylinder_lst sphere_lst plan_lst \
 					reflections light light2 light3 tuple_init scene scene2 scene3 matrix_transformation matrix_transformation2\
 					camera shadows plan colors cylinder cylinder2 cone cone2 cone_lst\
